create role

+++ src/controllers/notice.ts
@@ -9,7 +9,7 @@ import {
     Delete,
     Params,
     Put,
-    Query
+    Query,
 } from '@decorators/express';
 import { Injectable } from '@decorators/di';
 import logger from '../modules/logger';
@@ -18,6 +18,9 @@ import { IWriteNoticeBody } from '../types/notice';
 import HttpStatusCode from '../constants/HttpStatusCode';
 import { celebrate } from 'celebrate';
 import { noticeWriteValidator } from '../validators/notice';
+import { accessTokenGuard } from '../modules/passport';
+import { ErrorHandler } from '../modules/ErrorHandler';
+import { ACCOUNT_TYPE, ROLE_FLAG } from '../constants';
 
 @Controller('/notice')
 @Injectable()
@@ -41,7 +44,7 @@ export class NoticeController {
             page: page,
             count: count,
             sort: sort,
-            search: search
+            search: search,
         });
         return res.status(HttpStatusCode.OK).json(result);
     }
@@ -53,37 +56,53 @@ export class NoticeController {
         @Params() id: number
     ) {
         const result = await this.service.get(id);
-        if(!result) return res.status(HttpStatusCode.BAD_REQUEST).send('result not found');
+        if (!result)
+            return res
+                .status(HttpStatusCode.BAD_REQUEST)
+                .send('result not found');
         return res.status(HttpStatusCode.OK).json(result);
     }
 
-    @Post('/')
+    @Post('/', [accessTokenGuard])
     async write(
         @Request() req: IRequest,
         @Response() res: IResponse,
         @Body() body: IWriteNoticeBody
     ) {
+        if (!req.user)
+            return ErrorHandler(new TypeError('req.user is undefined'), res);
+        if (req.user.role & ROLE_FLAG.admin)
+            return res.status(HttpStatusCode.UNAUTHORIZED);
         const result = await this.service.write(body);
         return res.status(HttpStatusCode.OK).json(result);
     }
 
-    @Put('/:id')
+    @Put('/:id', [accessTokenGuard])
     async update(
         @Request() req: IRequest,
-        @Response() res : IResponse,
+        @Response() res: IResponse,
         @Params('id') id: number,
         @Body() body: IWriteNoticeBody
     ) {
-        const result = await this.service.update(id,body);
+        if (!req.user)
+            return ErrorHandler(new TypeError('req.user is undefined'), res);
+        if (req.user.role & ROLE_FLAG.admin)
+            return res.status(HttpStatusCode.UNAUTHORIZED);
+        const result = await this.service.update(id, body);
         return res.status(HttpStatusCode.OK).json(result);
     }
 
-    @Delete('/:id')
+    @Delete('/:id', [accessTokenGuard])
     async delete(
         @Request() req: IRequest,
         @Response() res: IResponse,
         @Params('id') id: number
     ) {
+        if (!req.user)
+            return ErrorHandler(new TypeError('req.user is undefined'), res);
+        if (req.user.role & ROLE_FLAG.admin) {
+            return res.status(HttpStatusCode.UNAUTHORIZED);
+        }
         const result = await this.service.delete(id);
         return res.status(HttpStatusCode.OK).json(result);
     }
